[FLINK-3229] Move KinesisSerializationSchema to org.apache.flink.streaming.connectors.kinesis.serialization package
